added windchillFactor.cpp
Had a lot fun and am pretty proud of this one, this program calculates the windchill factor given the wind speed and temperature, /weather/ in Fahrenheit or Celcius :^)

Originally it was going to only print the Factor in Fahrenheit like the practice problem asked but I thought it was too simple and having it display only a number was just too vague to understand. So, i went out of my way to make it display both fahrenheit and celcius, and tell the user about safety, risk of frostbite and risk time. 

This helped me think more thoroughly about program structure, variables (both to use and to pass) and error checking