added triangleType.cpp
Focuses User-Defined Simple Data Types using Enum.
TriangleType.cpp figures out whether the length of each side of a triangle given by the user is either an equilateral, isosceles, scalene or not a triangle.

With this program, I learned to be more careful with checking for input failure with cin in a loop. I used a do.. while loop to check if the user has entered both input failure or a negative number as a length and if so, it would prompt the user of the issue and have the user input the sides again. One issue I found using gdb was that the user can enter the first 2 sides as normal but the 3rd side as a character and enter input failure, since it checks and clears the failure inside the loop, cin.clear() sets the last variable entered before the fail state to '0', and since the last variable was side3, the value of side3 will now be set to 0. Therefore when the program checks the loop condition, it would exit the loop because all sides are non negative numbers and finish the program with the triangle type as 'No Triangle'! 
To prevent the loop from prematurely exiting, after the input failure has been cleared, the program resets each value of Sides back to -1, so the user can input each length correctly