Added testScores.cpp
Focuses on the use of structs. It reads the names and test scores of an amount of students (in this program 5 to keep it short but can be modified to a desired amount) and finds the grade of each student according to their test score. It finally displays a table of each students name, test score and grade, the highest test score, and who among the class have the highest score